python/requirements.txt
pandas>=1.5.0
requests>=2.28.0
urllib3>=1.26.0
pymysql>=1.0.0
opencv-python>=4.5.0
pyzbar>=0.1.9
qrcode>=7.3.1
Pillow>=8.0.0
